mathml/relations/html5-tree/autofocus.html

<!DOCTYPE html>
<script src="/resources/testharness.js" nonce="abc"></script>
<script src="/resources/testharnessreport.js" nonce="abc"></script>
<style>
    mrow { background: red; }
   :focus { background: blue;  }</style>
<body>
    <p>The box should be blue</p>
    <math>
      <mrow id="link" href="#target" tabindex="0" autofocus>
        <mspace id="space" width="150px" height="150px" style="background: transparent;"/>
      </mrow>
    </math>


    <script>
        async_test(t => {
          window.addEventListener('load', t.step_func_done(_ => { 
            assert_equals(document.activeElement, link);
          }));
        }, "a mrow with an href can be autofocused");
    </script>
</body>
</html>
